libraries/python/assistant-data-gen/pyproject.toml
[project]
name = "assistant_data_gen"
version = "0.1.0"
description = "Assistant Data Generation"
authors = [{ name = "Semantic Workbench Team" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "assistant-extensions==0.1.0",
    "chat-context-toolkit>=0.1.0",
    "pydantic>=2.10,<3.0",
    "pydantic-ai[openai,athropic]>=0.2,<0.3",
    "python-dotenv>=1.0.1,<2.0",
    "python-liquid>=2.0,<3.0",
    "PyYAML>=6.0,<7.0",
    "tiktoken>=0.9.0",
]

[tool.uv.sources]
assistant-extensions = { path = "../assistant-extensions", editable = true }
chat-context-toolkit = { path = "../chat-context-toolkit", editable = true }
openai-client = { path = "../openai-client", editable = true }

[dependency-groups]
dev = ["pyright>=1.1.401"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

